refactored route folders, removed warning by adding a key prop to ChampionBox in ChampionsList
