+++ client/src/components/ChampionsList/component.js
@@ -31,7 +31,7 @@ export default class ChampionsList extends Component {
                 Object.keys(champions).map(championName => {
                   if (championName !== 'version') {
                     var champion = champions[championName];
-                    return <ChampionBox champ={champion}/>;
+                    return <ChampionBox key={championName} champ={champion}/>;
                   }
                 })
                 )
